[VideoDB] Move VideoDatabase Version Upgrade code to separate .cpp Exact copy + clang-format
